examples/yolo/requirements.txt
numpy
opencv-python-headless
Pillow
scipy
